fix(CaseExpr,Generics, Predicate): desugaring CaseExpr, add support for Generics as match term of CaseExpr, add equals function for Predicate
CaseExpr: a CaseExpr occurring as expression in a statement is desugared as program variable whose value is assigned before the mentioned statement. Nested occurrences of CaseExpr are desugared recursively.

Generics: if a generic occurs as match term of CaseExpr with UnderscorePattern WildCardVar are generated. In this case, if no precondition is provided the solver CVC4 is not able to end the computation, while z3 succeeds anyway.

Predicate: the function equals for Predicate was missing, leading to problems with hard-coded matches between Predicate objects.